+++ .phan/config.php
@@ -314,6 +314,7 @@ return [
         'plugins/codeclimate/engine',
         'tool/make_stubs',
         'internal/dump_fallback_ast.php',
+        'internal/dump_html_styles.php',
         'internal/extract_arg_info.php',
         'internal/internalsignatures.php',
         'internal/package.php',